refact/SV-204: 예외처리 파일 위치 변경
