+++ src/main/java/ureca/nolmung/business/bookmark/BookmarkService.java
@@ -8,12 +8,12 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import ureca.nolmung.business.bookmark.request.BookmarkServiceRequest;
 import ureca.nolmung.business.bookmark.response.BookmarkResponse;
-import ureca.nolmung.exception.bookmark.BookmarkException;
-import ureca.nolmung.exception.bookmark.BookmarkExceptionType;
-import ureca.nolmung.exception.place.PlaceException;
-import ureca.nolmung.exception.place.PlaceExceptionType;
-import ureca.nolmung.exception.user.UserException;
-import ureca.nolmung.exception.user.UserExceptionType;
+import ureca.nolmung.implementation.bookmark.BookmarkException;
+import ureca.nolmung.implementation.bookmark.BookmarkExceptionType;
+import ureca.nolmung.implementation.place.PlaceException;
+import ureca.nolmung.implementation.place.PlaceExceptionType;
+import ureca.nolmung.implementation.user.UserException;
+import ureca.nolmung.implementation.user.UserExceptionType;
 import ureca.nolmung.implementation.bookmark.BookmarkManager;
 import ureca.nolmung.jpa.bookmark.Bookmark;
 import ureca.nolmung.jpa.place.Enum.Category;
