+++ src/main/java/ureca/nolmung/implementation/bookmark/BookmarkExceptionType.java
@@ -1,4 +1,4 @@
-package ureca.nolmung.exception.bookmark;
+package ureca.nolmung.implementation.bookmark;
 
 import lombok.AllArgsConstructor;
 import ureca.nolmung.exception.ExceptionType;
