+++ src/main/java/ureca/nolmung/implementation/review/ReviewManager.java
@@ -3,12 +3,10 @@ package ureca.nolmung.implementation.review;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import ureca.nolmung.business.review.dto.request.AddReviewReq;
-import ureca.nolmung.exception.label.LabelException;
-import ureca.nolmung.exception.label.LabelExceptionType;
-import ureca.nolmung.exception.place.PlaceException;
-import ureca.nolmung.exception.place.PlaceExceptionType;
-import ureca.nolmung.exception.review.ReviewException;
-import ureca.nolmung.exception.review.ReviewExceptionType;
+import ureca.nolmung.implementation.label.LabelException;
+import ureca.nolmung.implementation.label.LabelExceptionType;
+import ureca.nolmung.implementation.place.PlaceException;
+import ureca.nolmung.implementation.place.PlaceExceptionType;
 import ureca.nolmung.jpa.label.Label;
 import ureca.nolmung.jpa.label.LabelId;
 import ureca.nolmung.jpa.place.Place;
