+++ src/main/java/ureca/nolmung/implementation/user/UserExceptionType.java
@@ -1,13 +1,13 @@
-package ureca.nolmung.exception.label;
+package ureca.nolmung.implementation.user;
 
 import lombok.AllArgsConstructor;
 import ureca.nolmung.exception.ExceptionType;
 import ureca.nolmung.exception.Status;
 
 @AllArgsConstructor
-public enum LabelExceptionType implements ExceptionType {
+public enum UserExceptionType implements ExceptionType {
 
-    LABEL_NOT_FOUND_EXCEPTION(Status.NOT_FOUND, "해당 라벨이 존재하지 않습니다.");
+    USER_NOT_FOUND_EXCEPTION(Status.NOT_FOUND, "회원이 존재하지 않습니다.");
 
     private final Status status;
     private final String message;
